requirements.txt
Flask>=3.0,<4.0
cryptography>=42.0
flask-limiter>=3.5
Pillow>=10.0
pytest>=8.0

